requirements.txt
# local package
-e .

# ArcGIS Pro 3.x base (arcgispro-py3) env packages
# -e C:\Program Files\ArcGIS\Pro\bin\Python\envs\arcgispro-py3

# Hugin Python API packages
# -e C:\Program Files\Hugin Expert\HUGIN 9.3 (x64)\HDE9.3Python\Lib 

# external requirements
# conda install -c conda-forge python-dotenv
python-dotenv
